Restructured files, style tweak
