--- src/components/ModelsBySolution/index.scss
@@ -1,24 +1,0 @@
-@use 'uswds-core' as *;
-
-.model-solution-card {
-    .usa-card__container {
-        border-radius: .25rem;
-        border: 2px solid color('base-lighter');    
-        height: auto !important;
-        margin: 0 !important;
-    }
-
-    &__title {
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-        max-width: 90%;
-    }
-
-    &__body-text {
-        display: -webkit-box;
-        -webkit-line-clamp: 2;
-        -webkit-box-orient: vertical;
-        overflow: hidden;
-    }
-}
