+++ src/views/HomeNew/components/ModelsBySolutions/card.tsx
