+++ src/views/HomeNew/components/ModelsBySolutions/index.scss
@@ -14,3 +14,26 @@
         flex-direction: row-reverse;
     }
 }
+
+.model-solution-card {
+    .usa-card__container {
+        border-radius: .25rem;
+        border: 2px solid color('base-lighter');    
+        height: auto !important;
+        margin: 0 !important;
+    }
+
+    &__title {
+        overflow: hidden;
+        white-space: nowrap;
+        text-overflow: ellipsis;
+        max-width: 90%;
+    }
+
+    &__body-text {
+        display: -webkit-box;
+        -webkit-line-clamp: 2;
+        -webkit-box-orient: vertical;
+        overflow: hidden;
+    }
+}
